feat: 추첨 메서드 도메인 변경 (#108)
