+++ src/main/java/com/hyundai/softeer/backend/domain/lottery/entity/LotteryEvent.java
@@ -1,16 +1,11 @@
 package com.hyundai.softeer.backend.domain.lottery.entity;
 
-import jakarta.persistence.Column;
-import jakarta.persistence.Id;
-import jakarta.persistence.MappedSuperclass;
+import com.hyundai.softeer.backend.domain.subevent.entity.SubEvent;
+import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 import lombok.experimental.SuperBuilder;
-import org.hibernate.annotations.JdbcTypeCode;
-import org.hibernate.type.SqlTypes;
-
-import java.util.Map;
 
 @MappedSuperclass
 @SuperBuilder
@@ -19,10 +14,10 @@ import java.util.Map;
 @NoArgsConstructor
 public class LotteryEvent {
     @Id
-    @Column(name = "sub_event_id", nullable = false)
-    private Long subEventId;
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
 
-    @JdbcTypeCode(SqlTypes.JSON)
-    @Column(columnDefinition = "json")
-    private Map<String, Object> winnersMeta;
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "sub_event_id")
+    private SubEvent subEvent;
 }
